moved haversine to Location class
