+++ classes/State.py
@@ -3,7 +3,6 @@ from classes.Cluster import Cluster
 from classes.Vehicle import Vehicle
 from classes.Action import Action
 from system_simulation.scripts import system_simulate
-from math import sqrt, pi, sin, cos, atan2
 import matplotlib.pyplot as plt
 
 from globals import GEOSPATIAL_BOUND_NEW
@@ -52,18 +51,8 @@ class State:
                 if cluster == neighbour:
                     neighbour_distance.append(0.0)
                 else:
-                    (cluster_center_lat, cluster_center_lon,) = cluster.get_location()
-                    (
-                        neighbour_center_lat,
-                        neighbour_center_lon,
-                    ) = neighbour.get_location()
                     neighbour_distance.append(
-                        State.haversine(
-                            cluster_center_lat,
-                            cluster_center_lon,
-                            neighbour_center_lat,
-                            neighbour_center_lon,
-                        )
+                        cluster.distance_to(*neighbour.get_location())
                     )
             distance_matrix.append(neighbour_distance)
         return distance_matrix
@@ -173,26 +162,6 @@ class State:
 
     def __repr__(self):
         return f"State: Current cluster={self.current_cluster}"
-
-    @staticmethod
-    def haversine(lat1, lon1, lat2, lon2):
-        """
-        Compute the distance between two points in meters
-        :param lat1: Coordinate 1 lat
-        :param lon1: Coordinate 1 lon
-        :param lat2: Coordinate 2 lat
-        :param lon2: Coordinate 2 lon
-        :return: Kilometers between coordinates
-        """
-        radius = 6378.137
-        d_lat = lat2 * pi / 180 - lat1 * pi / 180
-        d_lon = lon2 * pi / 180 - lon1 * pi / 180
-        a = sin(d_lat / 2) * sin(d_lat / 2) + cos(lat1 * pi / 180) * cos(
-            lat2 * pi / 180
-        ) * sin(d_lon / 2) * sin(d_lon / 2)
-        c = 2 * atan2(sqrt(a), sqrt(1 - a))
-        distance = radius * c
-        return distance
 
     def visualize_clustering(self):
         fig, ax = plt.subplots(figsize=[10, 6])
